+++ libs/libc/stream/lib_stdinstream.c
@@ -60,12 +60,12 @@ static int stdinstream_getc(FAR struct lib_instream_s *self)
  * Name: stdinstream_gets
  ****************************************************************************/
 
-static int stdinstream_gets(FAR struct lib_instream_s *self,
-                            FAR void *buffer, int len)
+static ssize_t stdinstream_gets(FAR struct lib_instream_s *self,
+                                FAR void *buffer, size_t len)
 {
   FAR struct lib_stdinstream_s *stream =
                                        (FAR struct lib_stdinstream_s *)self;
-  int nread = 0;
+  ssize_t nread = 0;
 
   DEBUGASSERT(self);
